Juniper-Medical-Chatbot-AI--master/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
groq==0.11.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
numpy
transformers
